+++ packages/building-rollup/modern-config.js
@@ -4,7 +4,6 @@ const { findSupportedBrowsers } = require('@open-wc/building-utils');
 const resolve = require('rollup-plugin-node-resolve');
 const { terser } = require('rollup-plugin-terser');
 const babel = require('rollup-plugin-babel');
-const minifyHTML = require('rollup-plugin-minify-html-literals').default;
 const modernWeb = require('./plugins/rollup-plugin-modern-web/rollup-plugin-modern-web.js');
 
 const production = !process.env.ROLLUP_WATCH;
@@ -24,12 +23,6 @@ module.exports = function createBasicConfig(_options) {
       sourcemap: true,
     },
     plugins: [
-      // minify html and css template literals if in production
-      production &&
-        minifyHTML({
-          failOnError: true,
-        }),
-
       // parse input index.html as input and feed any modules found to rollup
       options.input.endsWith('.html') && modernWeb(),
 
@@ -44,7 +37,21 @@ module.exports = function createBasicConfig(_options) {
           // rollup rewrites import.meta.url, but makes them point to the file location after bundling
           // we want the location before bundling
           'bundled-import-meta',
-        ],
+          production && [
+            'template-html-minifier',
+            {
+              modules: {
+                'lit-html': ['html'],
+                'lit-element': ['html', { name: 'css', encapsulation: 'style' }],
+              },
+              htmlMinifier: {
+                collapseWhitespace: true,
+                removeComments: true,
+                caseSensitive: true,
+              },
+            },
+          ],
+        ].filter(_ => !!_),
         presets: [
           [
             '@babel/env',