+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.3](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.12.2...@open-wc/testing@0.12.3) (2019-05-14)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [0.12.2](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.12.1...@open-wc/testing@0.12.2) (2019-05-14)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
 ## [0.12.1](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.12.0...@open-wc/testing@0.12.1) (2019-05-06)
 
 **Note:** Version bump only for package @open-wc/testing